CompressiblePacket is needed
In someone's infinite wisdom, they felt it nessecary to not only deprecate
the CompressablePacket interface but made a new interface with the correct
spelling and ignored the deprecation warnings.

Bravo.

I've added back the correct interface with the correct spelling.

Lastly, PacketAlert is now PacketShowAchievement to better show what this
does.

Signed-off-by: Chris Sanders <zidane@outlook.com>